# api/docs/swagger.yaml
basePath: /
definitions:
  EditAccountRequest:
    properties:
      action:
        example: removeAvatar
        type: string
      mimeType:
        example: image/png
        type: string
    required:
    - action
    type: object
  ErrorResponse:
    properties:
      code:
        example: InternalError
        type: string
      error:
        example: An unexpected error occurred
        type: string
    type: object
  Exercise:
    properties:
      asset:
        $ref: '#/definitions/ImageDescription'
      category:
        example: Body weight
        type: string
      instructions:
        example: Keep your body straight and lower yourself until your chest almost
          touches the ground.
        type: string
      name:
        example: Push Up
        type: string
      target:
        example: Chest
        type: string
      thumbnail:
        $ref: '#/definitions/ImageDescription'
    type: object
  ExercisesResponse:
    properties:
      exercises:
        items:
          $ref: '#/definitions/Exercise'
        type: array
    type: object
  FeedbackRequest:
    properties:
      message:
        example: Good job!
        type: string
    required:
    - message
    type: object
  ImageDescription:
    properties:
      height:
        example: 100
        type: integer
      link:
        example: https://example.com/image.jpg
        type: string
      width:
        example: 100
        type: integer
    type: object
  PresignedUrlResponse:
    properties:
      fields:
        additionalProperties:
          type: string
        type: object
      url:
        type: string
    type: object
  Set:
    properties:
      completed:
        example: true
        type: boolean
      distance:
        example: 10
        type: number
      duration:
        example: 10
        type: number
      id:
        example: "2025-07-18T05:40:48.329406Z"
        type: string
      reps:
        example: 10
        type: integer
      weight:
        example: 100
        type: number
    required:
    - completed
    - id
    type: object
  Template:
    properties:
      exercises:
        items:
          $ref: '#/definitions/TemplateExercise'
        type: array
      id:
        example: "2"
        type: string
      name:
        example: Legs & Shoulders
        type: string
      order:
        example: 1
        type: integer
    type: object
  TemplateExercise:
    properties:
      exercise:
        type: string
      id:
        example: "2025-07-18T05:40:48.329406Z"
        type: string
      order:
        type: integer
      sets:
        items:
          $ref: '#/definitions/Set'
        type: array
    type: object
  TemplateIn:
    properties:
      exercises:
        items:
          $ref: '#/definitions/TemplateExercise'
        type: array
      id:
        type: string
      name:
        type: string
      order:
        type: integer
    type: object
  User:
    properties:
      accountDeletionSchedule:
        example: arn:aws:scheduler:ca-central-1:123:schedule/account-deletions/account-deletion-123
        type: string
      avatar:
        example: https://example.com/avatar.png
        type: string
      displayName:
        example: jane_doe
        type: string
      email:
        example: jane_doe@mail.com
        type: string
      id:
        example: HW4beTVvbTUPRxun9MXZxwKPjmC2
        type: string
      scheduledForDeletionAt:
        example: "2022-01-01T00:00:00.000Z"
        type: string
    required:
    - displayName
    - email
    - id
    type: object
  Workout:
    properties:
      end:
        example: "2023-01-01T12:00:00Z"
        type: string
      exercises:
        items:
          $ref: '#/definitions/WorkoutExercise'
        type: array
      id:
        example: "2025-07-18T05:40:48.329406Z"
        type: string
      name:
        example: Legs
        type: string
      start:
        example: "2023-01-01T12:00:00Z"
        type: string
    type: object
  WorkoutExercise:
    properties:
      exercise:
        example: Push Up
        type: string
      id:
        example: "2025-07-18T05:40:48.329406Z"
        type: string
      sets:
        items:
          $ref: '#/definitions/Set'
        type: array
    type: object
  WorkoutExerciseIn:
    properties:
      exercise:
        example: Push Up
        type: string
      id:
        example: "2025-07-18T05:40:48.329406Z"
        type: string
      order:
        example: 1
        type: integer
      sets:
        items:
          $ref: '#/definitions/Set'
        type: array
    required:
    - exercise
    - id
    type: object
  WorkoutIn:
    properties:
      end:
        example: "2023-01-01T12:00:00Z"
        type: string
      exercises:
        items:
          $ref: '#/definitions/WorkoutExerciseIn'
        type: array
      id:
        example: "2025-07-18T05:40:48.329406Z"
        type: string
      name:
        example: Legs
        type: string
      start:
        example: "2023-01-01T12:00:00Z"
        type: string
    required:
    - exercises
    - id
    - start
    type: object
  WorkoutResponse:
    properties:
      cursor:
        type: string
      workouts:
        items:
          $ref: '#/definitions/Workout'
        type: array
    type: object
host: localhost:8080
info:
  contact:
    name: Kit
    url: https://github.com/kit-g
  description: A simple fitness tracker API
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Heart of Yours API
  version: "1.0"
paths:
  /accounts:
    delete:
      consumes:
      - application/json
      description: Schedules account deletion for the authenticated user
      operationId: deleteAccount
      parameters:
      - description: Account ID
        in: path
        name: accountId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete user account
      tags:
      - accounts
    post:
      consumes:
      - application/json
      description: Accounts are managed by Firebase so we just need to store them
      operationId: registerAccount
      parameters:
      - description: User request
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/User'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/User'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Creates an account record
      tags:
      - accounts
  /accounts/{accountId}:
    get:
      consumes:
      - application/json
      description: Returns user account information for the authenticated user
      operationId: getAccount
      parameters:
      - description: Account ID
        in: path
        name: accountId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/User'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user account
      tags:
      - accounts
    put:
      consumes:
      - application/json
      description: 'Performs various account editing actions: undoAccountDeletion,
        removeAvatar, uploadAvatar'
      operationId: editAccount
      parameters:
      - description: Account ID
        in: path
        name: accountId
        required: true
        type: string
      - description: Edit account request
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/EditAccountRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/PresignedUrlResponse'
        "204":
          description: No Content
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Edit user account
      tags:
      - accounts
  /exercises:
    get:
      consumes:
      - application/json
      description: Returns all exercises in a single page
      operationId: getExercises
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ExercisesResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: List all exercises
      tags:
      - workouts
  /feedback:
    post:
      consumes:
      - application/json
      description: Allows users to submit feedback with a message and optionally attach
        a screenshot
      parameters:
      - description: Feedback details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/FeedbackRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/PresignedUrlResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Submit user feedback
      tags:
      - feedback
  /templates:
    get:
      consumes:
      - application/json
      description: Returns all workout templates for the authenticated user
      operationId: getTemplates
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/Template'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Lists workout templates
      tags:
      - templates
    post:
      consumes:
      - application/json
      description: Validates, saves and returns a workout template
      operationId: makeTemplate
      parameters:
      - description: Template request
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/TemplateIn'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Template'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Creates a workout template
      tags:
      - templates
  /templates/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a specific workout template by ID
      operationId: deleteTemplate
      parameters:
      - description: Template ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete workout template
      tags:
      - templates
    get:
      consumes:
      - application/json
      description: Returns a specific workout template by ID
      operationId: getTemplate
      parameters:
      - description: Template ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Template'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get workout template
      tags:
      - templates
  /workouts:
    get:
      consumes:
      - application/json
      description: Returns paginated list of user workouts with exercises and sets
      operationId: getWorkouts
      parameters:
      - description: Page size for pagination
        in: query
        name: pageSize
        type: integer
      - description: Cursor for pagination
        in: query
        name: cursor
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/WorkoutResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Returns user workouts
      tags:
      - workouts
    post:
      consumes:
      - application/json
      description: Validates, saves and returns a workout
      operationId: makeWorkout
      parameters:
      - description: Workout request
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/WorkoutIn'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Workout'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Creates a workout
      tags:
      - workouts
  /workouts/{workoutId}:
    delete:
      consumes:
      - application/json
      description: Deletes a workout by ID
      operationId: deleteWorkout
      parameters:
      - description: Workout ID
        in: path
        name: workoutId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Deletes a workout
      tags:
      - workouts
    get:
      consumes:
      - application/json
      description: Returns a single workout by ID with exercises and sets
      operationId: getWorkout
      parameters:
      - description: Workout ID
        in: path
        name: workoutId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Workout'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Returns a workout
      tags:
      - workouts
swagger: "2.0"